Organized program files into a programs folder
