+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/NKNOpMode_AdvancedMovementTest.java
@@ -6,8 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.AbsoluteMovementTestProgram;
-import org.nknsd.robotics.team.FlowSensorTestProgram;
+import org.nknsd.robotics.team.programs.AbsoluteMovementTestProgram;
 
 import java.util.concurrent.TimeUnit;
