+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/NKNOpMode_PotProgram.java
@@ -6,8 +6,7 @@ import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.nknsd.robotics.framework.NKNProgram;
-import org.nknsd.robotics.team.FlowSensorTestProgram;
-import org.nknsd.robotics.team.PotentiometerProgram;
+import org.nknsd.robotics.team.programs.PotentiometerProgram;
 
 import java.util.concurrent.TimeUnit;
